requirements.txt
# Core RAG dependencies
anthropic>=0.40.0
chromadb>=0.4.0

# Document processing
pypdf>=4.0.0
python-docx>=1.0.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Testing
pytest>=8.0.0

sentence-transformers>=2.2.0